+++ src/main/java/shopping/infrastructure/dto/ExchangeRateResponse.java
@@ -1,0 +1,14 @@
+package shopping.infrastructure.dto;
+
+public class ExchangeRateResponse {
+
+    private Double rate;
+
+    public ExchangeRateResponse(Double rate) {
+        this.rate = rate;
+    }
+
+    public Double getRate() {
+        return rate;
+    }
+}
