+++ src/test/java/shopping/domain/auth/BearerExtractorTest.java
@@ -1,4 +1,4 @@
-package shopping.auth;
+package shopping.domain.auth;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
