+++ src/test/java/shopping/domain/order/OrderTest.java
@@ -6,7 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatCode;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import shopping.exchange.MockExchangeRateApi;
+import shopping.infrastructure.MockExchangeRateApi;
 import shopping.global.vo.Name;
 import shopping.global.vo.Price;
 import shopping.global.vo.Quantity;
