+++ src/test/java/shopping/infrastructure/ExchangeRateApiTest.java
@@ -1,4 +1,4 @@
-package shopping.exchange;
+package shopping.infrastructure;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.offset;
@@ -12,7 +12,7 @@ public class ExchangeRateApiTest {
     @DisplayName("exchangeRateApi 를 통해 환율을 얻을 수 있다.")
     void api_이용하여_환율_얻음(){
         // given
-        Double actual = new MockExchangeRateApi().callExchangeRate();
+        Double actual = new MockExchangeRateApi().callExchangeRate().getRate();
         // when & then
         assertThat(actual).isCloseTo(1300.1, offset(0.01));
         assertThat(actual).isNotCloseTo(1300.2,offset(0.01));
