+++ src/test/java/shopping/infrastructure/MockExchangeRateApi.java
@@ -1,15 +1,16 @@
-package shopping.exchange;
+package shopping.infrastructure;
 
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 import shopping.infrastructure.ExchangeRateApi;
+import shopping.infrastructure.dto.ExchangeRateResponse;
 
-@Profile("test")
 @Component
+@Profile("test")
 public class MockExchangeRateApi implements ExchangeRateApi {
 
     @Override
-    public Double callExchangeRate() {
-        return 1300.1;
+    public ExchangeRateResponse callExchangeRate() {
+        return new ExchangeRateResponse(1300.1);
     }
 }
